+++ R/block_cluster_two_arm_designer.R
@@ -7,7 +7,7 @@
 #' Units are assigned to treatment using complete block cluster random assignment. Treatment effects can be specified either by providing \code{control_mean} and \code{treatment_mean}
 #' or by specifying an \code{ate}. Estimation uses differences in means accounting for blocks and clusters.
 #' 
-#' Total N is given by \code{N_blocks*N_clusters_in_block*N_i_in_cluster}.
+#' In the usual case \code{N} is not provided by the user but is determined by \code{N_blocks, N_clusters_in_block, N_i_in_cluster} (when these are integers \code{N} is the product of these three numbers).
 #' 
 #' Normal shocks can be specified at the individual, cluster, and block levels. If individual level shocks are not specified and cluster and block 
 #' level variances sum to less than 1, then individual level shocks are set such that total variance in outcomes equals 1.
@@ -16,15 +16,17 @@
 #' 
 #' See \href{https://declaredesign.org/library/articles/block_cluster_two_arm.html}{vignette online}.
 #' 
+#' @param N An integer. Total number of units. Usually not specified as \code{N} is determined by \code{N_blocks}, \code{N_clusters_in_block}, and \code{N_i_in_cluster}. If \code{N_blocks}, and  \code{N_clusters_in_block}, and \code{N_i_in_cluster} are specified then \code{N} is overridden. If these are not specified and \code{N} is specified then designer attempts to guess sizes of levels to approximate \code{N}, with preference for a design without blocks or clusters. 
 #' @param N_blocks An integer. Number of blocks. Defaults to 1 for no blocks. 
-#' @param N_clusters_in_block An integer. Number of clusters in each block. This is the total \code{N} when \code{N_blocks} and \code{N_i_in_cluster} are at default values. 
-#' @param N_i_in_cluster An integer. Individuals per cluster. Defaults to 1 for no clusters.
+#' @param N_clusters_in_block An integer or vector of integers of length \code{N_blocks}. Number of clusters in each block. This is the total \code{N} when \code{N_blocks} and \code{N_i_in_cluster} are at default values. 
+#' @param N_i_in_cluster An integer or vector of integers of length \code{sum(N_clusters_in_block)}. Individuals per cluster. Defaults to 1 for no clusters.
+#' @param sd A nonnegative number. Overall standard deviation (combining individual level, cluster level, and block level shocks). Defaults to 1. Overridden if incompatible with other user-specified shocks. 
 #' @param sd_block A nonnegative number. Standard deviation of block level shocks.
 #' @param sd_cluster A nonnegative number. Standard deviation of cluster level shock.
-#' @param sd_i_0 A nonnegative number. Standard deviation of individual level shock in control. For small \code{sd_block} and \code{sd_cluster}, \code{sd_i_0} defaults to make total variance = 1.
+#' @param sd_i_0 A nonnegative number. Standard deviation of individual level shock in control. For sufficiently small \code{sd_block} and \code{sd_cluster}, \code{sd_i_0} defaults to make total variance = sd.
 #' @param sd_i_1 A nonnegative number. Standard deviation of individual level shock in treatment. Defaults to \code{sd_i_0}.
 #' @param rho A number in [-1,1]. Correlation in individual shock between potential outcomes for treatment and control.
-#' @param prob A number in (0,1). Treatment assignment probability.
+#' @param assignment_prob A number in (0,1). Treatment assignment probability.
 #' @param control_mean A number. Average outcome in control.
 #' @param ate A number. Average treatment effect. Alternative to specifying \code{treatment_mean}. Note that \code{ate} is an argument for the designer but it does not appear as an argument in design code (design code uses \code{control_mean} and \code{treatment_mean} only).
 #' @param treatment_mean A number. Average outcome in treatment. If \code{treatment_mean} is not provided then it is calculated as \code{control_mean + ate}. If both \code{ate} and  \code{treatment_mean} are provided then only  \code{treatment_mean} is used. 
@@ -32,37 +34,57 @@
 #' @author \href{https://declaredesign.org/}{DeclareDesign Team}
 #' @concept experiment 
 #' @concept blocking
-#' @import DeclareDesign stats utils fabricatr estimatr randomizr
+#' @importFrom DeclareDesign declare_assignment declare_estimand declare_estimator declare_population declare_potential_outcomes declare_reveal
+#' @importFrom fabricatr fabricate add_level
+#' @importFrom randomizr conduct_ra 
+#' @importFrom estimatr tidy lm_robust
+#' @importFrom rlang is_integerish
+#' @importFrom stats rnorm
 #' @export
 #' @examples
 #' # Generate a design using default arguments:
 #' block_cluster_two_arm_design <- block_cluster_two_arm_designer()
-#' 
+#' block_cluster_uneven <- block_cluster_two_arm_designer(
+#'        N_blocks = 3, N_clusters_in_block = 2:4, N_i_in_cluster = 1:9)
+#' # A design in which number of clusters of cluster size is not specified
+#' # but N and block size are:        
+#' block_cluster_guess <- block_cluster_two_arm_designer(N = 24, N_blocks = 3)
+#' # A design in which cluster size is not specified but N and block size are 
+#' # and target N is not met returns an error:        
+#' \dontrun{block_cluster_guess_2 <- block_cluster_two_arm_designer(N = 24,
+#' N_blocks = 3, N_clusters_in_block = 3)}
 #'
 
-block_cluster_two_arm_designer <- function(N_blocks = 1,
-                                           N_clusters_in_block = 100,
-                                           N_i_in_cluster = 1,
+block_cluster_two_arm_designer <- function(N = NULL,
+                                           N_blocks = 1,
+                                           N_clusters_in_block = ifelse(is.null(N), 100, round(N/N_blocks)),
+                                           N_i_in_cluster = ifelse(is.null(N), 1, round(N/mean(N_blocks*N_clusters_in_block))),
+                                           sd = 1,
                                            sd_block = .5,
                                            sd_cluster = .5,
-                                           sd_i_0 = sqrt(max(0, 1 - sd_block^2 - sd_cluster^2)),
+                                           sd_i_0 = sqrt(max(0, sd^2 - sd_block^2 - sd_cluster^2)),
                                            sd_i_1 = sd_i_0,
                                            rho = 1,
-                                           prob = .5,
+                                           assignment_prob = .5,
                                            control_mean = 0,
                                            ate = 0,
                                            treatment_mean = control_mean + ate
 ){  
+
   if(any(N_blocks < 1, N_clusters_in_block < 1, N_i_in_cluster < 1) ||
-     any(!rlang::is_integerish(N_blocks), 
-         !rlang::is_integerish(N_clusters_in_block), 
-         !rlang::is_integerish(N_i_in_cluster))) stop("N_* arguments must be positive integers")
+     any(!is_integerish(N_blocks), 
+         !is_integerish(N_clusters_in_block), 
+         !is_integerish(N_i_in_cluster))) stop("N_* arguments must be positive integers")
   if(sd_block < 0) stop("sd_block must be nonnegative")
   if(sd_cluster < 0) stop("sd_cluster must be nonnegative")
   if(sd_i_0 < 0) stop("sd_i_0 must be nonnegative")
   if(sd_i_1 < 0) stop("sd_i_1 must be nonnegative")
-  if(prob<= 0 || prob >= 1) stop("prob must be in (0,1)")
+  if(assignment_prob<= 0 || assignment_prob >= 1) stop("assignment_prob must be in (0,1)")
   if(rho< -1 || rho > 1) stop("correlation must be in [-1,1]")
+  if(!is.null(N)) {design_N <- ifelse(length(N_i_in_cluster)>1, sum(N_i_in_cluster), sum(N_i_in_cluster*N_blocks*N_clusters_in_block))
+                  if(N != design_N) stop(paste0("The design N of ", design_N, " is inconsistent with the user specified N of ", N, 
+                                                  ". Likely due to integer problems in specified block or cluster sizes. Better to fully specify N for each level and not provide an argument for overal N.")
+                                            )}
   {{{    
     # M: Model
     population <- declare_population(
@@ -81,13 +103,13 @@ block_cluster_two_arm_designer <- function(N_blocks = 1,
     
     potentials <- declare_potential_outcomes(
       Y ~ (1 - Z) * (control_mean    + u_0*sd_i_0 + u_b + u_c) + 
-        Z *       (treatment_mean  + u_1*sd_i_1 + u_b + u_c) )
+           Z *      (treatment_mean  + u_1*sd_i_1 + u_b + u_c) )
     
     # I: Inquiry
     estimand <- declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0))
     
     # D: Data Strategy
-    assignment <- declare_assignment(prob = prob, blocks = blocks, clusters = clusters)
+    assignment <- declare_assignment(prob = assignment_prob, blocks = blocks, clusters = clusters)
     reveal     <- declare_reveal()
     
     # A: Answer Strategy
@@ -106,7 +128,7 @@ block_cluster_two_arm_designer <- function(N_blocks = 1,
   
   attr(block_cluster_two_arm_design, "code") <- 
     construct_design_code(block_cluster_two_arm_designer, match.call.defaults(),
-                          exclude_args = "ate",
+                          exclude_args = c("ate", "sd", "N"),
                           arguments_as_values = TRUE)
   
   block_cluster_two_arm_design