+++ R/process_tracing_designer.R
@@ -34,7 +34,13 @@
 #' @author \href{https://declaredesign.org/}{DeclareDesign Team}
 #' @concept qualitative 
 #' @concept process tracing
-#' @import DeclareDesign stats utils fabricatr estimatr randomizr
+#' @importFrom DeclareDesign declare_diagnosands declare_estimand declare_estimator declare_population declare_sampling declare_step diagnose_design draw_data get_estimands get_estimates set_diagnosands
+#' @importFrom fabricatr fabricate fabricate
+#' @importFrom randomizr conduct_ra draw_rs 
+#' @importFrom estimatr tidy
+#' @importFrom stats rbinom
+#' @importFrom rlang is_integerish is_character
+#' @importFrom utils data
 #' @export
 #' @examples
 #' # Generate a process-tracing design using default arguments:
@@ -77,7 +83,7 @@ process_tracing_designer <- function(
   label_E1 = "Smoking Gun",
   label_E2 = "Straw in the Wind"
 ){
-  if(!rlang::is_integerish(N) || N < 1) stop("N must be a positive integer.")
+  if(!is_integerish(N) || N < 1) stop("N must be a positive integer.")
   if(prob_X < 0 || prob_X > 1) stop("prob_X must be in [0,1].")
   if(length(process_proportions) != 4) stop("process_proportions must be of length 4.")
   if(sum(process_proportions) != 1 || any(process_proportions < 0) || any(process_proportions> 1)) stop("process_proportions must be a 3-simplex.")
@@ -93,8 +99,8 @@ process_tracing_designer <- function(
     c((1 - p1) * (1 - p2) + r,p2 * (1 - p1) - r, p1 * (1 - p2) - r, p1 * p2 + r)}
   if(min(test_prob(p_E1_H, p_E2_H, cor_E1E2_H)) < 0) stop("Correlation coefficient not compatible with probabilities")
   if(min(test_prob(p_E1_not_H, p_E2_not_H, cor_E1E2_not_H)) < 0) stop("Correlation coefficient not compatible with probabilities")
-  if(!rlang::is_character(label_E1) || length(label_E1) > 1) stop("label_E1 must be a character of length 1.")
-  if(!rlang::is_character(label_E2) || length(label_E2) > 1) stop("label_E2 must be a character of length 1.")
+  if(!is_character(label_E1) || length(label_E1) > 1) stop("label_E1 must be a character of length 1.")
+  if(!is_character(label_E2) || length(label_E2) > 1) stop("label_E2 must be a character of length 1.")
   {{{
     # M: Model
     population <- declare_population(
@@ -158,25 +164,25 @@ process_tracing_designer <- function(
     )
     smoking_gun <- declare_estimator(
       handler = bayes_estimator,
-      p_E_H = ifelse(data$E1, p_E1_H, 1 - p_E1_H),
-      p_E_not_H = ifelse(data$E1, p_E1_not_H, 1 - p_E1_not_H),
+      p_E_H = with(data, ifelse(E1, p_E1_H, 1 - p_E1_H)),
+      p_E_not_H = with(data, ifelse(E1, p_E1_not_H, 1 - p_E1_not_H)),
       label = label_E1,
-      result = data$E1
+      result = with(data, E1)
     )
     straw_in_wind <- declare_estimator(
       handler = bayes_estimator,
-      p_E_H = ifelse(data$E2, p_E2_H, 1 - p_E2_H),
-      p_E_not_H = ifelse(data$E2, p_E2_not_H, 1 - p_E2_not_H),
+      p_E_H = with(data, ifelse(E2, p_E2_H, 1 - p_E2_H)),
+      p_E_not_H = with(data, ifelse(E2, p_E2_not_H, 1 - p_E2_not_H)),
       label = label_E2,
-      result = data$E2
+      result = with(data, E2)
     )
     
     joint_test <- declare_estimator(
       handler = bayes_estimator,
-      p_E_H = joint_prob_H[c("00", "01", "10", "11") %in% data$test_results],
-      p_E_not_H = joint_prob_not_H[c("00", "01", "10", "11") %in% data$test_results],
+      p_E_H = joint_prob_H[c("00", "01", "10", "11") %in% data[["test_results"]]],
+      p_E_not_H = joint_prob_not_H[c("00", "01", "10", "11") %in% data[["test_results"]]],
       label = paste(label_E1, "and", label_E2),
-      result = data$test_results
+      result = data[["test_results"]]
     )
     
     # Design
@@ -211,8 +217,8 @@ attr(process_tracing_designer,"shiny_arguments") <- list(
 )
 attr(process_tracing_designer,"tips") <- c(
   prior_H = "Prior probability that the hypothesis that X causes Y is true.",
-  p_E1_H = "Probability of observing the first piece of evidence given X causes Y.",
-  cor_E1E2_H = "Correlation in first and second piece of evidence given X causes Y."
+  p_E1_H = "Probability of observing the first piece of evidence given X indeed causes Y.",
+  cor_E1E2_H = "Correlation in first and second pieces of evidence given X indeed causes Y."
 )
 attr(process_tracing_designer,"description") <- "A process-tracing design in which two pieces of evidence are sought and used to update about whether X caused Y using Bayes' rule."