+++ R/randomized_response_designer.R
@@ -15,9 +15,11 @@
 #' @author \href{https://declaredesign.org/}{DeclareDesign Team}
 #' @concept experiment
 #' @concept descriptive
-#' @import DeclareDesign stats utils fabricatr estimatr randomizr
+#' @importFrom DeclareDesign declare_assignment declare_diagnosands declare_estimand declare_estimator declare_population declare_potential_outcomes declare_reveal set_diagnosands tidy_estimator
+#' @importFrom fabricatr fabricate draw_binary
+#' @importFrom randomizr conduct_ra 
+#' @importFrom estimatr tidy 
 #' @export
-#'
 #' @examples
 #' # Generate a randomized response design using default arguments:
 #' randomized_response_design <- randomized_response_designer()