+++ R/simple_two_arm_designer.R
@@ -10,17 +10,21 @@
 #' See \href{https://declaredesign.org/library/articles/simple_two_arm.html}{vignette online}.
 #' 
 #' @param N An integer. Sample size.
-#' @param prob A number in [0,1]. Probability of assignment to treatment.
+#' @param assignment_prob A number in [0,1]. Probability of assignment to treatment.
 #' @param control_mean A number. Average outcome in control.
 #' @param control_sd A positive number. Standard deviation in control.
 #' @param ate A number. Average treatment effect.
 #' @param treatment_mean A number. Average outcome in treatment. Overrides \code{ate} if both specified.
-#' @param treatment_sd  A non-negative number. Standard deviation in treatment. By default equals \code{control_sd}.
+#' @param treatment_sd  A nonnegative number. Standard deviation in treatment. By default equals \code{control_sd}.
 #' @param rho A number in [-1,1]. Correlation between treatment and control outcomes.
 #' @return A simple two-arm design.
 #' @author \href{https://declaredesign.org/}{DeclareDesign Team}
 #' @concept experiment
-#' @import DeclareDesign stats utils fabricatr estimatr randomizr
+#' @importFrom DeclareDesign declare_assignment declare_estimand declare_estimator declare_population declare_potential_outcomes declare_reveal
+#' @importFrom fabricatr fabricate 
+#' @importFrom randomizr conduct_ra 
+#' @importFrom estimatr tidy
+#' @importFrom stats rnorm
 #' @export
 #'
 #' @examples
@@ -29,7 +33,7 @@
 
 
 simple_two_arm_designer <- function(N = 100,
-                                    prob = .5,
+                                    assignment_prob = .5,
                                     control_mean = 0,
                                     control_sd = 1,
                                     ate = 1,
@@ -38,7 +42,7 @@ simple_two_arm_designer <- function(N = 100,
                                     rho = 1
 ){
   if(control_sd < 0 ) stop("control_sd must be non-negative")
-  if(prob < 0 || prob > 1) stop("prob must be in [0,1]")
+  if(assignment_prob < 0 || assignment_prob > 1) stop("assignment_prob must be in [0,1]")
   if(abs(rho) > 1) stop("rho must be in [-1,1]")
   {{{
     # M: Model
@@ -55,11 +59,11 @@ simple_two_arm_designer <- function(N = 100,
     estimand <- declare_estimand(ATE = mean(Y_Z_1 - Y_Z_0))
     
     # D: Data Strategy
-    assignment <- declare_assignment(prob = prob)
+    assignment <- declare_assignment(prob = assignment_prob)
+    reveal_Y    <- declare_reveal()
     
     # A: Answer Strategy
     estimator <- declare_estimator(Y ~ Z, estimand = estimand)
-    reveal_Y    <- declare_reveal()
     
     # Design
     simple_two_arm_design <- population + potentials + estimand + assignment + reveal_Y + estimator
@@ -83,7 +87,7 @@ attr(simple_two_arm_designer, "tips") <-
   )
 
 attr(simple_two_arm_designer, "description") <- "
-<p> A simple two arm design of sample size <code>N</code> and with constant average treatment effect equal to <code>ate</code>.
+<p> A simple two arm design of sample size <code>N</code> and with average treatment effect equal to <code>ate</code>.
 "